Struct language_detector::English

pub struct English<'a> {
    // some fields omitted
}

This is the English detector structure.

It requires to be initialized to be build the internal trigram mapping to score.

Examples

To initialize the structure and check if a phrase is in English do the following:

use language_detector::English;

let detection = English::new();
assert!(detection.is_english("The king and the queen"));
assert!(!detection.is_english("O rei e a rainha"));

Methods

impl<'a> English<'a>

fn new() -> English<'a>

fn is_english(&self, phrase: &str) -> bool